requirements.txt
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pushbullet.py==0.12.0
pycodestyle==2.10.0
pyjokes==0.6.0
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.13
websocket-client==1.4.2
word2number==1.1
